+++ GCSViews/ConfigurationView/ConfigHWRangeFinder.cs
@@ -8,8 +8,8 @@ namespace MissionPlanner.GCSViews.ConfigurationView
     public partial class ConfigHWRangeFinder : UserControl, IActivate, IDeactivate
     {
         bool startup = true;
-        private const float rad2deg = (float) (180/Math.PI);
-        private const float deg2rad = (float) (1.0/rad2deg);
+        private const double rad2deg = (float) (180/Math.PI);
+        private const double deg2rad = (float) (1.0/rad2deg);
 
         public ConfigHWRangeFinder()
         {
